// quizonline-frontend/src/app/api/generated/model/course-enrollment-status-enum.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `pending` - Pending approval * `active` - Active * `completed` - Completed * `cancelled` - Cancelled
 */
export enum CourseEnrollmentStatusEnumDto {

    Pending = 'pending',

    Active = 'active',

    Completed = 'completed',

    Cancelled = 'cancelled'
}